fix: navbar on mobile and Home Page overflow
