+++ src/components/Navbar/index.css
@@ -1,0 +1,67 @@
+/* NavBar Section  */
+.nav {
+  background-color: var(--bg-nav);
+  user-select: none;
+  position: sticky;
+  top: 0;
+  z-index: 999;
+}
+.nav .container {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-between;
+  margin-inline: auto;
+  padding-block: 5px;
+}
+.div--logo {
+  height: 80px;
+  width: 80px;
+}
+.div--logo img {
+  height: 80px;
+  width: 80px;
+}
+.nav-items {
+  display: flex;
+  align-items: center;
+  gap: 3vw;
+  color: var(--primary);
+}
+.nav-item {
+  position: relative;
+  font-size: 20px;
+  list-style: none;
+  font-weight: 600;
+  color: inherit;
+}
+.nav-item::after,
+.nav-item span::after {
+  content: "";
+  height: 2px;
+  background-color: var(--primary);
+  position: absolute;
+  right: 0;
+  bottom: 0;
+  width: 0;
+  transition: all 200ms cubic-bezier(0.4, 0, 0.2, 1);
+}
+.nav-item span.active::after,
+.nav-item:hover::after {
+  width: 100%;
+  left: 0;
+  right: auto;
+}
+.nav-item,
+.nav-item:active,
+.nav-item:visited {
+  text-decoration: none;
+}
+@media only screen and (max-width: 768px) {
+  .nav .container {
+    flex-direction: column;
+  }
+  .nav-items {
+    padding-block: 3px;
+  }
+}
