+++ src/pages/About/index.css
@@ -1,0 +1,113 @@
+/* About Section */
+.about--section {
+  margin-inline: auto;
+  display: flex;
+  flex-grow: 1;
+  height: 100%;
+  align-items: center;
+  flex-direction: column-reverse;
+  justify-content: center;
+  gap: 3rem;
+}
+.about--title {
+  /* font-family: monospace; */
+  font-size: 2rem;
+  color: var(--primary);
+  font-style: italic;
+}
+
+.about--content {
+  margin-top: 1rem;
+  font-size: 1rem;
+}
+.about--para {
+  margin-top: 1rem;
+}
+.about--btn {
+  -webkit-appearance: button;
+  -moz-appearance: button;
+  appearance: button;
+  margin-top: 2rem;
+  height: 3rem;
+  width: 10rem;
+  background-color: var(--primary);
+  font-size: 1em;
+  border: none;
+  border-radius: 2px;
+  color: var(--bg-color);
+  font-weight: 800;
+  border-radius: 9999px;
+}
+.about--btn:hover {
+  background-color: transparent;
+  border: 3px solid var(--primary);
+  cursor: pointer;
+}
+.about--profile {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+.about--profile img {
+  border-radius: 9999px;
+  height: 200px;
+  width: 200px;
+  box-shadow: inset 0 0 50px #fff, inset 20px 0 80px #f0f,
+    inset -20px 0 80px #0ff, inset 20px 0 300px #f0f, inset -20px 0 300px #0ff,
+    0 0 50px #fff, -10px 0 80px #f0f, 10px 0 80px #0ff;
+}
+
+@media only screen and (max-width: 768px) {
+  /* HOME PAGE */
+  /* .about--section {
+    flex-direction: column-reverse;
+    padding-block-start: 4rem;
+    padding-block-end: 2rem;
+    gap: 3rem;
+  }
+  .about--title {
+    font-size: 2rem;
+  }
+  .about--content {
+    font-size: 1.2rem;
+  } */
+}
+@media (min-width: 640px) {
+  .about--profile img {
+    height: 300px;
+    width: 300px;
+  }
+}
+@media (min-width: 768px) {
+}
+@media (min-width: 1024px) {
+  .about--section {
+    gap: 5rem;
+    flex-direction: row;
+    text-align: left;
+  }
+
+  .about--title {
+    font-size: 3rem;
+  }
+  .about--content {
+    font-size: 1.5rem;
+  }
+  .about--profile {
+    width: 50%;
+  }
+  .about--profile img {
+    height: 400px;
+    width: 400px;
+  }
+}
+@media (min-width: 1280px) {
+  .about--title {
+    font-size: 4rem;
+  }
+  .about--content {
+    font-size: 2.5rem;
+  }
+}
+@media (min-width: 1536px) {
+}
